Reorganize fuzz and integration test code
This moves the TestRtc and associated functions used
in the integration tests to the module str0m::_internal_text_exports.
By having it in the main source tree we can share the code between
fuzz tests and integration tests.
