+++ _str0m_test/src/lib.rs
@@ -1,0 +1,2 @@
+//! This can be empty since feature flags are additive, and the _internal_test_exports
+//! is enabled for str0m in the main compilation unit.
