+++ fuzz/Cargo.toml
@@ -12,6 +12,7 @@ libfuzzer-sys = "0.4"
 
 [dependencies.str0m]
 path = ".."
+features = ["_internal_test_exports"]
 
 # Prevent this from interfering with workspaces
 [workspace]
