+++ src/_internal_test_exports/fuzz.rs
@@ -1,0 +1,29 @@
+//! Exported fuzz targets to get them part of the compilation with feature `_internal_test_exports`.
+
+use std::time::Duration;
+use std::time::Instant;
+
+use crate::streams::rtx_cache_buf::EvictingBuffer;
+
+pub fn rtx_buffer(data: &[u8]) {
+    if data.len() < 4 {
+        return;
+    }
+
+    let buf_size = u16::from_be_bytes([data[0], data[1]]);
+    let max_age = data[2] as u64;
+    let max_size = data[3] as usize;
+    let mut buf = EvictingBuffer::new(buf_size as usize, Duration::from_secs(max_age), max_size);
+    let mut now = Instant::now();
+    let mut pos = 0;
+
+    for d in &data[4..] {
+        now += Duration::from_millis(*d as u64);
+        if d % 2 == 0 {
+            buf.maybe_evict(now)
+        } else {
+            pos += *d as u64;
+            buf.push(pos, now, d);
+        }
+    }
+}
